update packages to last versions; switch to PostgreSQL; separate mixins from models
