+++ .gitignore
@@ -1,3 +1,4 @@
+env/
 venv/
 .env
 .flaskenv
@@ -14,3 +15,5 @@ htmlcov/
 dist/
 build/
 *.egg-info/
+
+.vs/
