+++ requirements.txt
@@ -1,38 +1,86 @@
-alembic==1.5.2
-aniso8601==8.1.0
-cachelib==0.1.1
-click==7.1.2
-DateTime==4.3
-dnspython==2.1.0
-email-validator==1.1.2
-Flask==1.1.2
-Flask-Admin==1.5.7
-Flask-Authorize==0.2.2
-Flask-Dropzone==1.5.4
-Flask-JSGlue==0.3.1
-Flask-Login==0.5.0
-flask-marshmallow==0.14.0
-Flask-Migrate==2.6.0
-Flask-RESTful==0.3.8
-Flask-Session==0.3.2
-Flask-SQLAlchemy==2.4.4
-Flask-Uploads==0.2.1
-Flask-WTF==0.14.3
-idna==3.1
-IPTCInfo3==2.1.4
-itsdangerous==1.1.0
-Jinja2==2.11.2
-Mako==1.1.4
-MarkupSafe==1.1.1
-marshmallow==3.10.0
-marshmallow-sqlalchemy==0.24.1
-Pillow==8.1.0
-python-dateutil==2.8.1
-python-dotenv==0.15.0
-python-editor==1.0.4
-pytz==2020.5
-six==1.15.0
-SQLAlchemy==1.3.22
-Werkzeug==1.0.1
-WTForms==2.3.3
-zope.interface==5.2.0
+a l e m b i c = = 1 . 7 . 6  
+ a n i s o 8 6 0 1 = = 9 . 0 . 1  
+ c a c h e l i b = = 0 . 6 . 0  
+ c l i c k = = 8 . 0 . 3  
+ c o l o r a m a = = 0 . 4 . 4  
+ D a t e T i m e = = 4 . 4  
+ d n s p y t h o n = = 2 . 2 . 0  
+ e m a i l - v a l i d a t o r = = 1 . 1 . 3  
+ F l a s k = = 2 . 0 . 2  
+ F l a s k - A d m i n = = 1 . 6 . 0  
+ F l a s k - A u t h o r i z e = = 0 . 2 . 6  
+ F l a s k - D r o p z o n e = = 1 . 6 . 0  
+ F l a s k - J S G l u e = = 0 . 3 . 1  
+ F l a s k - L o g i n = = 0 . 5 . 0  
+ f l a s k - m a r s h m a l l o w = = 0 . 1 4 . 0  
+ F l a s k - M i g r a t e = = 3 . 1 . 0  
+ F l a s k - R E S T f u l = = 0 . 3 . 9  
+ F l a s k - S Q L A l c h e m y = = 2 . 5 . 1  
+ F l a s k - U p l o a d s = = 0 . 2 . 1  
+ F l a s k - W T F = = 1 . 0 . 0  
+ g r e e n l e t = = 1 . 1 . 2  
+ i d n a = = 3 . 3  
+ i m p o r t l i b - m e t a d a t a = = 4 . 1 1 . 0  
+ i m p o r t l i b - r e s o u r c e s = = 5 . 4 . 0  
+ I P T C I n f o 3 = = 2 . 1 . 4  
+ i t s d a n g e r o u s = = 2 . 0 . 1  
+ J i n j a 2 = = 3 . 0 . 3  
+ M a k o = = 1 . 1 . 4  
+ M a r k u p S a f e = = 2 . 0 . 1  
+ m a r s h m a l l o w = = 3 . 1 4 . 1  
+ m a r s h m a l l o w - s q l a l c h e m y = = 0 . 2 7 . 0  
+ P i l l o w = = 9 . 0 . 1  
+ p y t h o n - d a t e u t i l = = 2 . 8 . 2  
+ p y t h o n - d o t e n v = = 0 . 1 9 . 2  
+ p y t h o n - e d i t o r = = 1 . 0 . 4  
+ p y t z = = 2 0 2 1 . 3  
+ s i x = = 1 . 1 6 . 0  
+ S Q L A l c h e m y = = 1 . 4 . 3 1  
+ W e r k z e u g = = 2 . 0 . 3  
+ W T F o r m s = = 3 . 0 . 1  
+ z i p p = = 3 . 7 . 0  
+ z o p e . i n t e r f a c e = = 5 . 5 . 0 . d e v 0  
+ a l e m b i c = = 1 . 7 . 6  
+ a n i s o 8 6 0 1 = = 9 . 0 . 1  
+ c a c h e l i b = = 0 . 6 . 0  
+ c l i c k = = 8 . 0 . 3  
+ c o l o r a m a = = 0 . 4 . 4  
+ D a t e T i m e = = 4 . 4  
+ d n s p y t h o n = = 2 . 2 . 0  
+ e m a i l - v a l i d a t o r = = 1 . 1 . 3  
+ F l a s k = = 2 . 0 . 2  
+ F l a s k - A d m i n = = 1 . 6 . 0  
+ F l a s k - A u t h o r i z e = = 0 . 2 . 6  
+ F l a s k - D r o p z o n e = = 1 . 6 . 0  
+ F l a s k - J S G l u e = = 0 . 3 . 1  
+ F l a s k - L o g i n = = 0 . 5 . 0  
+ f l a s k - m a r s h m a l l o w = = 0 . 1 4 . 0  
+ F l a s k - M i g r a t e = = 3 . 1 . 0  
+ F l a s k - R E S T f u l = = 0 . 3 . 9  
+ F l a s k - S Q L A l c h e m y = = 2 . 5 . 1  
+ F l a s k - U p l o a d s = = 0 . 2 . 1  
+ F l a s k - W T F = = 1 . 0 . 0  
+ g r e e n l e t = = 1 . 1 . 2  
+ i d n a = = 3 . 3  
+ i m p o r t l i b - m e t a d a t a = = 4 . 1 1 . 0  
+ i m p o r t l i b - r e s o u r c e s = = 5 . 4 . 0  
+ I P T C I n f o 3 = = 2 . 1 . 4  
+ i t s d a n g e r o u s = = 2 . 0 . 1  
+ J i n j a 2 = = 3 . 0 . 3  
+ M a k o = = 1 . 1 . 4  
+ M a r k u p S a f e = = 2 . 0 . 1  
+ m a r s h m a l l o w = = 3 . 1 4 . 1  
+ m a r s h m a l l o w - s q l a l c h e m y = = 0 . 2 7 . 0  
+ P i l l o w = = 9 . 0 . 1  
+ p s y c o p g 2 = = 2 . 9 . 3  
+ p y t h o n - d a t e u t i l = = 2 . 8 . 2  
+ p y t h o n - d o t e n v = = 0 . 1 9 . 2  
+ p y t h o n - e d i t o r = = 1 . 0 . 4  
+ p y t z = = 2 0 2 1 . 3  
+ s i x = = 1 . 1 6 . 0  
+ S Q L A l c h e m y = = 1 . 4 . 3 1  
+ W e r k z e u g = = 2 . 0 . 3  
+ W T F o r m s = = 3 . 0 . 1  
+ z i p p = = 3 . 7 . 0  
+ z o p e . i n t e r f a c e = = 5 . 5 . 0 . d e v 0  
+
